chore: extract provider integration tests into tools/conformance
Move the three `//go:build integration` suites that lived under
sdkx/ into a dedicated, off-workspace `tools/conformance` module
pinned to released sdk v0.2.4 + sdkx v0.2.2. Bump sdkx's own sdk
require to v0.2.4 along the way.

Why
- The build-tag suites never compiled in CI (no job passed
  `-tags=integration`), so any breaking sdkx API change went
  unnoticed until someone ran them manually.
- They sat inside sdkx but were really developer tools requiring
  live provider credentials, not part of the SDK surface.
- `sdkx/internal/testenv` only existed to support them.

What changed
- New module `tools/conformance` (off-workspace, GOWORK=off, pinned
  to released sdk/sdkx — same pattern as bench/ and
  examples/voice-pipeline) hosting:
    llm/        provider_test.go + memory_test.go (shared providers
                matrix; same `package llm_test`)
    embedding/  embedding_test.go
    internal/testenv/  consolidated .env loader
- Build tags dropped; suites already had `t.Skip` when their env
  vars are unset, so a credential-less run is a clean no-op.
- Makefile: add tools/conformance to MODULES_OFFWORK + new
  `make conformance` target as the documented entry point.
- sdkx/go.mod: sdk v0.2.3 -> v0.2.4.
- .gitignore: allow tools/** (and prune some stale entries).
- Delete sdkx/internal/testenv and the three sdkx integration test
  files.

Effect
- `make test` now also runs the conformance suites under GOWORK=off
  against the pinned sdk/sdkx — without credentials they Skip, but
  the compile path is exercised every time, replacing the silent
  rot the build tag used to allow.
- Conformance suites now test the exact bytes downstream consumers
  pull, not unreleased workspace source.

Made-with: Cursor
